Moved types into source file
